service/requirements.txt
# Any requirements listed here will be symlinked to /lib and deployed to
# appspot.  See requirements_dev.txt for local-only requirements.

uritemplate
Werkzeug
httplib2
Flask
Flask-Cache
Flask-Login
httpagentparser
webob
itsdangerous
git+git://github.com/tylertreat/furious.git#egg=furious
PyGithub
